backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain_community
langchain-openai
langchain-chroma
chromadb
unstructured
pdfminer.six
streamlit
fpdf2
requests
python-docx
PyPDF2
pytesseract
Pillow
python-multipart
